backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
jinja2
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
fer
opencv-python
reportlab
python-dotenv
aiofiles
Pillow
torch
torchvision
torchaudio
numpy
moviepy<2.0.0
psycopg2-binary
boto3
